requirements.txt
numpy==1.15.0
spacy==2.0.12
stanford-corenlp
torch==0.4.1